requirements.txt
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
flask==3.0.3
